+++ packages/jarl-react/tool/safeJsonStringify.js
@@ -1,9 +1,7 @@
-"use strict";
-
 Object.defineProperty(exports, "__esModule", {
     value: true
 });
-var safeJsonStringify = function safeJsonStringify(o) {
+const safeJsonStringify = function safeJsonStringify(o) {
     try {
         return JSON.stringify(o);
     } catch (e) {